Ok.... May have broken everything for Shaurya, but updated everything to work with the latest rosjava version, fixed a large number of bugs, added back autonomy switching, and things seem to miraculously still be working fine in my tests.
Don't permanently add jars to the project anymore.  There is a new Maven based 
downloader and things get confusing with both.  rosmake should generate an eclipse
project that has the correct build path, if it doesn't, just edit the eclipse project
until it works, but DO NOT add or modify the svn files.